create two files js
